requirements.txt
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-polardb==1.8.1
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.2.1
DBUtils==1.4
idna==2.10
jmespath==0.10.0
pycparser==2.20
PyMySQL==0.10.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
